Keep single file - single class
