+++ src/Horarium.Test/TypeName/FirstNonGenericJob.cs
@@ -1,9 +1,4 @@
-﻿using Horarium.Attributes;
-
-namespace Horarium.Test.TypeName
+﻿namespace Horarium.Test.TypeName
 {
     public class FirstNonGenericJob { }
-
-    [GenericJob]
-    public class FirstNonGenericJobWithAttribute {}
 }
