+++ src/Horarium.Test/TypeName/GenericJob.cs
@@ -1,9 +1,4 @@
-using Horarium.Attributes;
-
 namespace Horarium.Test.TypeName
 {
     public class GenericJob<TFirst> { }
-
-    [GenericJob]
-    public class GenericJobWithAttribute<TFirst> { }
 }
