+++ src/Horarium.Test/TypeName/GenericJobWithTwoArguments.cs
@@ -1,9 +1,4 @@
-using Horarium.Attributes;
-
 namespace Horarium.Test.TypeName
 {
     public class GenericJobWithTwoArguments<TFirst, TSecond> { }
-
-    [GenericJob]
-    public class GenericJobWithTwoArgumentsWithAttribute<TFirst, TSecond> { }
 }
