+++ src/Horarium.Test/TypeName/GenericJobWithTwoArgumentsWithAttribute.cs
@@ -1,0 +1,7 @@
+﻿using Horarium.Attributes;
+
+namespace Horarium.Test.TypeName
+{
+    [GenericJob]
+    public class GenericJobWithTwoArgumentsWithAttribute<TFirst, TSecond> { }
+}
