+++ src/Horarium.Test/TypeName/SecondNonGenericJob.cs
@@ -1,9 +1,4 @@
-using Horarium.Attributes;
-
 namespace Horarium.Test.TypeName
 {
     public class SecondNonGenericJob { }
-
-    [GenericJob]
-    public class SecondNonGenericJobWithAttribute { }
 }
