+++ src/Horarium.Test/TypeName/SecondNonGenericJobWithAttribute.cs
@@ -1,0 +1,7 @@
+﻿using Horarium.Attributes;
+
+namespace Horarium.Test.TypeName
+{
+    [GenericJob]
+    public class SecondNonGenericJobWithAttribute { }
+}
